Add echo route

+++ src/routes/api/debug/clear/+server.ts
@@ -1,10 +1,11 @@
+import { dev } from '$app/environment';
 import { getDataDir } from '$lib/server/data/dataDir.js';
-import { json } from '@sveltejs/kit';
+import { error, json } from '@sveltejs/kit';
 import { rimraf } from 'rimraf';
 
 
 export async function DELETE({ request, cookies }) {
-
+  if (!dev) throw error(404);
   // Delete data directory
   await rimraf(getDataDir());
